chore(elasticsearch): Update Client (#25)
Tried to pull in Critical Security vulnerabilities but was unable to
build the package as the elasticsearch npm package has been deprecated
and is no longer available.

This PR updates to the available client, which is mostly a drop in
replacement and switches to use another type of connector which works
with the update elastic search client.

I was able to test and deploy this to the `dev` cluster but it does not
look like we have created the index or loaded it with data so I cannot
test end to end. I can see it gets the requests and attempts to search
but the index does not contain the data in dev.